fix(helm): avoid unnecessary Helm builds
Previously, Helm modules would get rebuilt ahead of deployment, even if
the service version hadn't changed (thus leading deploy tasks to not
redeploy the service, making the build step unnecessary).

We address that here by making the Helm build handler a no-op, and
instead running its logic inside the deployment handler.
