+++ core/src/plugins/kubernetes/helm/deployment.ts
@@ -7,10 +7,6 @@
  */
 
 import Bluebird from "bluebird"
-import { waitForResources } from "../status/status"
-import { helm } from "./helm-cli"
-import { HelmModule } from "./config"
-import { getChartPath, getReleaseName, getChartResources, getValueArgs, getBaseModule } from "./common"
 import {
   getReleaseStatus,
   HelmServiceStatus,
@@ -18,6 +14,18 @@ import {
   getPausedResources,
   gardenCloudAECPauseAnnotation,
 } from "./status"
+import tmp from "tmp-promise"
+import { waitForResources } from "../status/status"
+import { helm } from "./helm-cli"
+import { HelmModule } from "./config"
+import {
+  getChartPath,
+  getReleaseName,
+  getChartResources,
+  getValueArgs,
+  getBaseModule,
+  containsBuildSource,
+} from "./common"
 import { HotReloadableResource } from "../hot-reload/hot-reload"
 import { apply, deleteResources } from "../kubectl"
 import { KubernetesPluginContext, ServiceResourceSpec } from "../config"
@@ -30,6 +38,12 @@ import { getModuleNamespace, getModuleNamespaceStatus } from "../namespace"
 import { getHotReloadSpec, configureHotReload, getHotReloadContainerName } from "../hot-reload/helpers"
 import { configureDevMode, startDevModeSync } from "../dev-mode"
 import { KubeApi } from "../api"
+import { move } from "fs-extra"
+import { basename, join } from "path"
+import { ConfigurationError } from "../../../exceptions"
+import { BuildModuleParams, BuildResult } from "../../../types/plugin/module/build"
+import { deline } from "../../../util/string"
+import { LogEntry } from "../../../logger/log-entry"
 
 export async function deployHelmService({
   ctx,
@@ -40,6 +54,7 @@ export async function deployHelmService({
   devMode,
   hotReload,
 }: DeployServiceParams<HelmModule>): Promise<HelmServiceStatus> {
+  await prepareHelmModule({ ctx, module, log })
   let hotReloadSpec: ContainerHotReloadSpec | null = null
   let serviceResourceSpec: ServiceResourceSpec | null = null
   let serviceResource: HotReloadableResource | null = null
@@ -210,4 +225,64 @@ export async function deleteService(params: DeleteServiceParams): Promise<HelmSe
   log.setSuccess("Service deleted")
 
   return { state: "missing", detail: { remoteResources: [] } }
+}
+
+/**
+ * Fetches the module's Helm chart and updates local Helm repos if necessary.
+ */
+export async function prepareHelmModule({ ctx, module, log }: BuildModuleParams<HelmModule>): Promise<BuildResult> {
+  const k8sCtx = <KubernetesPluginContext>ctx
+  const baseModule = getBaseModule(module)
+
+  if (!baseModule && !(await containsBuildSource(module))) {
+    if (!module.spec.chart) {
+      throw new ConfigurationError(
+        deline`Module '${module.name}' neither specifies a chart name, base module,
+        nor contains chart sources at \`chartPath\`.`,
+        { module }
+      )
+    }
+    log.debug("Fetching chart...")
+    try {
+      await pullChart(k8sCtx, log, module)
+    } catch {
+      // Update the local helm repos and retry
+      log.debug("Updating Helm repos...")
+      // The stable repo is no longer added by default
+      await helm({
+        ctx: k8sCtx,
+        log,
+        args: ["repo", "add", "stable", "https://charts.helm.sh/stable", "--force-update"],
+      })
+      await helm({ ctx: k8sCtx, log, args: ["repo", "update"] })
+      log.debug("Fetching chart (after updating)...")
+      await pullChart(k8sCtx, log, module)
+    }
+  }
+
+  return { fresh: true }
+}
+
+async function pullChart(ctx: KubernetesPluginContext, log: LogEntry, module: HelmModule) {
+  const chartPath = await getChartPath(module)
+  const chartDir = basename(chartPath)
+
+  const tmpDir = await tmp.dir({ unsafeCleanup: true })
+
+  try {
+    const args = ["pull", module.spec.chart!, "--untar", "--untardir", tmpDir.path]
+
+    if (module.spec.version) {
+      args.push("--version", module.spec.version)
+    }
+    if (module.spec.repo) {
+      args.push("--repo", module.spec.repo)
+    }
+
+    await helm({ ctx, log, args: [...args], cwd: tmpDir.path })
+
+    await move(join(tmpDir.path, chartDir), chartPath, { overwrite: true })
+  } finally {
+    await tmpDir.cleanup()
+  }
 }
